Feature/msgs ros1 and ros2 compile (#185)
This PR updates this repo to allow for all the required autoware_msgs to compile in ROS1 and ROS2.  Due to the nature of these changes the change set it extensive, however, all changes to cpp files cover only variable name changes to make both the message sets compatible with both ROS1 and ROS2. 

* WIP make jsk_common_msgs compile in both ros1 and ros2

* Missed change on last commit

* WIP

* jsk_common_msgs builds in ros2

* Fix cmake version

* Buildable autoware_msgs

* All required autoware messages build in ROS2