Feature/wm intersection (#196)
Supports PR: usdot-fhwa-stol/carma-platform#1586
Major changes:

CarmaTrafficSignal can return stop_line of a specific entry_lanelet. This is because single signal_group can control multiple lanes and even opposite directions.
CarmaTrafficSignal's predictState also returns the end_time of current state. This will prove useful for user's to know when the signal will change. Also it is used to shift and modify the timers when new cycle info is received through SPAT.
OverwriteParameterVisitor and overwriteWithMatchingId. These functions are used to overwrite the parameter of regulatoryElement. it is used to match the memory address of received parameter to the on the already exists in the map with same id.
Related Issue
usdot-fhwa-stol/carma-platform#1587

Motivation and Context
This is to supprot TSMO Use Case 2.

How Has This Been Tested?
Local integration tested with real world MAP message.
At the time of this PR, faulty MAP msg converter is blocking full vehicle integration test.
usdot-fhwa-stol/carma-platform#1588