Fix/jsk common msgs build (#201)
This PR updates the lanelet2_extension and hardcoded_params to use ament_cmake instead of catkin when building in ROS2. Previously, this had been avoided to try to preserve backwards compatibility with carma-streets. However, it occurred to me that catkin could be made the default allowing that system to still work without ament_cmake being available. This fix also resolves a runtime linker error encountered for the lanelet2_extension package in the points_map_filter node. Because the lanelet2_extension library was compiled as a static rather than shared library the registration call was being optimized away due to lack of use. Making the library shared and marking the registration objects as static (specific to compilation unit). The registration is able to properly occur. Additionally, the lanelet2_extension package being compiled as c++17 instead of 14 when lanelet2 itself was compiled as c++14 was causing issues with the registration calls access to the RuleName parameter since the compilation of static constexpr changes in c++17. This only posed an issue with the registration itself. Downstream packages using lanelet2_extension may still be c++17 so long as they do not perform their own regulatory element registration.

I am currently leaving the unit test porting for future work as they can still be run with catkin and this task is otherwise holding up ROS2 development.

Related Issue
Supports usdot-fhwa-stol/carma-platform#1557